feat: update 计算模块资源增加状态统计条
