+++ containers/Compute/views/baremetal/index.vue
@@ -8,7 +8,11 @@
       </div>
     </page-header>
     <page-body>
-      <baremetal-list :id="listId" :cloud-env="cloudEnv" @refresh="refreshHandle" />
+      <baremetal-list
+        :id="listId"
+        :cloud-env="cloudEnv"
+        :filterParams="filterParams"
+        @refresh="refreshHandle" />
     </page-body>
   </div>
 </template>
@@ -16,6 +20,7 @@
 <script>
 import ResStatusTab from '@/sections/ResStatusTab'
 import BaremetalList from './components/List'
+import ResStatisticsMixin from '@/mixins/resStatisticsMixin'
 
 export default {
   name: 'BaremetalIndex',
@@ -23,6 +28,7 @@ export default {
     BaremetalList,
     ResStatusTab,
   },
+  mixins: [ResStatisticsMixin],
   data () {
     return {
       listId: 'BaremetalList',
@@ -30,66 +36,14 @@ export default {
         { key: '', label: this.$t('compute.text_4') },
       ],
       cloudEnv: '',
-      statusOpts: [],
-      statusArr: [],
-      errorFilterStatus: [],
-      otherFilterStatus: [],
+      resStaticsResource: 'servers',
     }
   },
   created () {
-    this.resStaticsManager = new this.$Manager('servers/statistics')
-    this.fetchResStatistics()
-  },
-  beforeDestroy () {
-    this.resStaticsManager = null
-  },
-  methods: {
-    fetchResStatistics () {
-      this.errorFilterStatus = []
-      this.otherFilterStatus = []
-
-      this.resStaticsManager.list({
-        params: {
-          scope: this.$store.getters.scope,
-          baremetal: true,
-        },
-      }).then(res => {
-        const { ready = {}, running = {} } = res.data
-
-        // 统计
-        let total = 0
-        let error = 0
-        let other = 0
-        for (const k in res.data) {
-          total += res.data[k].total_count
-          if (new RegExp('fail').test(k)) {
-            this.errorFilterStatus.push(k)
-            error += res.data[k].total_count
-          } else {
-            if (!['running', 'ready'].includes(k)) {
-              this.otherFilterStatus.push(k)
-              other += res.data[k].total_count
-            }
-          }
-        }
-
-        const statusOpts = [
-          { title: this.$t('compute.text_576'), type: 'total', num: total },
-          { title: this.$t('compute.text_574'), type: 'running', num: running?.total_count || 0 },
-          { title: this.$t('compute.text_273'), type: 'ready', num: ready?.total_count || 0 },
-          { title: this.$t('common_623', [this.$t('scope.text_61')]), type: 'error', num: error },
-          { title: this.$t('compute.text_674'), type: 'other', num: other },
-        ]
-        this.statusOpts = statusOpts
-        this.statusArr = Object.keys(res.data)
-      }).catch(err => {
-        console.error(err)
-        this.statusOpts = []
-      })
-    },
-    refreshHandle () {
-      this.fetchResStatistics()
-    },
+    this.fetchResStatistics({
+      scope: this.$store.getters.scope,
+      baremetal: true,
+    })
   },
 }
 </script>
